requirements.txt
# HomeCentralMaid Requirements

# Core dependencies
requests>=2.31.0
ollama>=0.1.0
pyyaml>=6.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.7.0

# Optional dependencies for future plugins
# psutil>=5.9.0  # For system_info plugin
